CONFLICT (content): Merge conflict in openmp/libomptarget/plugins/level0/CMakeLists.txt

  commit 2e0cd2c3268212013a49292abc2da6c8b5f5ead3
  Author: Yilong Guo <vfirst218@gmail.com>
  Date:   Thu May 19 11:26:28 2022 +0800

      Revert "[OpenMP][Offload][L0] Enable sycl wrapper for interop (#4976)" (#5111)

      This reverts commit 9d6bc3664a3aed9e37568dc736adb948f9db8b45.

+++ openmp/libomptarget/include/rtl.h
@@ -79,7 +79,7 @@ struct RTLInfoTy {
   typedef void (deinit_ty)(void);
 #if INTEL_CUSTOMIZATION
   typedef __tgt_interop *(create_interop_ty)(int32_t, int32_t, int32_t,
-                                             int32_t *);
+                                             intptr_t *);
   typedef int32_t(release_interop_ty)(int32_t, __tgt_interop *);
   typedef int32_t(use_interop_ty)(int32_t, __tgt_interop *);
   typedef int32_t(get_num_interop_properties_ty)(int32_t);
